# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TradingBot API
  version: 0.1.0
  description: |
    Full API and WebSocket reference for the TradingBot Go backend.
    Base path for all REST endpoints: `/api/v1`
    WebSocket endpoints are documented under the `x-websocket` extension sections at the bottom.

    ## Authentication
    All protected endpoints require a Bearer JWT in the `Authorization` header.
    Tokens are obtained via `POST /api/v1/auth/login` or `POST /api/v1/auth/signup`.
    Access tokens expire in **15 minutes**. Use `POST /api/v1/auth/refresh` to rotate.

    ## Response Envelope
    Every response (success or error) is wrapped in a standard envelope:
    ```json
    { "data": {}, "meta": { "requestId": "", "serverTime": "", "version": "v1" } }
    { "error": { "code": "", "message": "", "details": {} }, "meta": { ... } }
    ```
servers:
  - url: http://localhost:8000
    description: Local development server
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Security Schemes
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        HS256-signed JWT. Obtain via /auth/login or /auth/signup.
        Include as: `Authorization: Bearer <token>`

  # ───────────────────────────────────────────────────────────────────────────
  # Shared Schemas
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
    Meta:
      type: object
      properties:
        requestId:
          type: string
          example: req_1234567890
        serverTime:
          type: string
          format: date-time
          example: "2026-05-06T23:10:00+05:30"
        version:
          type: string
          example: v1
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          example: 50
        cursor:
          type: string
        nextCursor:
          type: string
        hasMore:
          type: boolean
    APIError:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: One or more fields are invalid.
        details:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/APIError'
        meta:
          $ref: '#/components/schemas/Meta'

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
          description: Rotated refresh token — old token is invalidated
        expiresIn:
          type: integer
          example: 900

    # ── Health ────────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: invalid or expired token
            meta:
              requestId: req_123
              version: v1
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue new access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RefreshResponse'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...